Fix type of `http.response.status_code` attribute

+++ tracing-opentelemetry-instrumentation-sdk/src/http/http_server.rs
@@ -37,7 +37,7 @@ pub fn make_span_from_request<B>(req: &http::Request<B>) -> tracing::Span {
 
 pub fn update_span_from_response<B>(span: &tracing::Span, response: &http::Response<B>) {
     let status = response.status();
-    span.record(HTTP_RESPONSE_STATUS_CODE, status.as_u16());
+    span.record(HTTP_RESPONSE_STATUS_CODE, i64::from(status.as_u16()));
 
     if status.is_server_error() {
         span.record(OTEL_STATUS_CODE, "ERROR");
